// src/app/types/WorkType.tsx
export interface IWorkType {
    id: number;
  name: string;
  image: string;
  muscleGroups: string[];
  equipment: string;
  difficulty: string;
  duration: number;        
  caloriesBurned: number;  
  sets: number;
  reps: string;            
  rating: number;      
  description: string;
  instructions: string[];
  }